fixed is inf bug in nps_fdm_gazebbo

+++ sw/simulator/nps/nps_fdm_gazebo.cpp
@@ -770,7 +770,7 @@ static void gazebo_read_video(void)
 
 }*/
 
- if (!isinf(ray_left->Range(0))&&!isinf(ray_right->Range(0))&&(ray_left->Range(0) + ray_right->Range(0))<3.0f){
+ if (!isinf((float)ray_left->Range(0)) && !isinf((float)ray_right->Range(0)) && (ray_left->Range(0) + ray_right->Range(0)) < 3.0f ){
  float corner_distance_compensation =1 - fabs(ray_left->Range(0) + ray_right->Range(0))/3;
  cout<<"corner compensation :"<<corner_distance_compensation<<endl;
  distance_closest_obstacle -= corner_distance_compensation;}
